+++ tests/test_pyopenssl.py
@@ -4,7 +4,7 @@ import pytest
 
 from OpenSSL.crypto import FILETYPE_PEM, load_certificate
 
-from service_identity._common import (
+from service_identity.common import (
     DNS_ID,
     DNSPattern,
     IPAddress_ID,
@@ -17,7 +17,7 @@ from service_identity.exceptions import (
     VerificationError,
 )
 from service_identity.pyopenssl import (
-    extract_ids,
+    extract_patterns,
     verify_hostname,
     verify_ip_address,
 )
@@ -97,24 +97,24 @@ class TestExtractIDs:
         """
         Returns the correct DNSPattern from a certificate.
         """
-        rv = extract_ids(CERT_DNS_ONLY)
+        rv = extract_patterns(CERT_DNS_ONLY)
         assert [
-            DNSPattern(b"www.twistedmatrix.com"),
-            DNSPattern(b"twistedmatrix.com"),
+            DNSPattern.from_bytes(b"www.twistedmatrix.com"),
+            DNSPattern.from_bytes(b"twistedmatrix.com"),
         ] == rv
 
     def test_cn_ids_are_ignored(self):
         """
         commonName is not supported anymore and therefore ignored.
         """
-        assert [] == extract_ids(CERT_CN_ONLY)
+        assert [] == extract_patterns(CERT_CN_ONLY)
 
     def test_uri(self):
         """
         Returns the correct URIPattern from a certificate.
         """
-        rv = extract_ids(CERT_OTHER_NAME)
-        assert [URIPattern(b"http://example.com/")] == [
+        rv = extract_patterns(CERT_OTHER_NAME)
+        assert [URIPattern.from_bytes(b"http://example.com/")] == [
             id for id in rv if isinstance(id, URIPattern)
         ]
 
@@ -122,13 +122,15 @@ class TestExtractIDs:
         """
         Returns IP patterns.
         """
-        rv = extract_ids(CERT_EVERYTHING)
+        rv = extract_patterns(CERT_EVERYTHING)
 
         assert [
-            DNSPattern(pattern=b"service.identity.invalid"),
-            DNSPattern(pattern=b"*.wildcard.service.identity.invalid"),
-            DNSPattern(pattern=b"service.identity.invalid"),
-            DNSPattern(pattern=b"single.service.identity.invalid"),
+            DNSPattern.from_bytes(pattern=b"service.identity.invalid"),
+            DNSPattern.from_bytes(
+                pattern=b"*.wildcard.service.identity.invalid"
+            ),
+            DNSPattern.from_bytes(pattern=b"service.identity.invalid"),
+            DNSPattern.from_bytes(pattern=b"single.service.identity.invalid"),
             IPAddressPattern(pattern=ipaddress.IPv4Address("1.1.1.1")),
             IPAddressPattern(pattern=ipaddress.IPv6Address("::1")),
             IPAddressPattern(pattern=ipaddress.IPv4Address("2.2.2.2")),